Plan Nutricional

+++ CapaEntidad/EntPlatillo.cs
@@ -10,9 +10,9 @@ namespace CapaEntidad
     {
         public int IdPlatillo { get; set; }
         public string Nombre { get; set; } = string.Empty;
-        public double Calorias { get; set; }
-        public double Carbohidratos { get; set; }
-        public double Grasas { get; set; }
-        public double Proteinas { get; set; }
+        public int Calorias { get; set; }
+        public decimal Carbohidratos { get; set; }
+        public decimal Grasas { get; set; }
+        public decimal Proteinas { get; set; }
     }
 }
